server/requirements.txt
Django
pymongo
djongo
django-mongodb-engine
djangorestframework
requests
